fix: nitpick cleanups from code review (N-23 through N-32)
- os module imported properly instead of __import__("os")
- browse_folder distinguishes a real COM/dialog failure from a user cancel
- fix Marshal.PtrToStringUni + leak-safe cleanup in the folder-picker COM call
- lock browse_folder so two concurrent clicks can't open two dialogs
- drop redundant int(round()) in SmartResolutionMultiplier
- unify "CraftKit —" docstring prefix across nodes
- comment OUTPUT_NODE=True on the status-label nodes so it isn't "cleaned up" later
- add .jfif/.avif/.heic to SmartBatchResize's supported extensions, document non-recursive scanning
- add a progress bar to SmartBatchResize (degrades gracefully outside a live ComfyUI)
- add readme/requires-python and a dependency-rationale comment to pyproject.toml

+++ smart_resolution_multiplier.py
@@ -38,8 +38,8 @@ class SmartResolutionMultiplier:
     def run(self, image, multiplier, multiple_of):
         h, w = image.shape[1], image.shape[2]
 
-        new_w = int(round(w * multiplier))
-        new_h = int(round(h * multiplier))
+        new_w = round(w * multiplier)
+        new_h = round(h * multiplier)
 
         if multiple_of > 1:
             new_w = max(multiple_of, round(new_w / multiple_of) * multiple_of)
